#73 migration fix

+++ database/migrations/3_sales_management_pipelines_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
         Schema::create(config('sales-management.tablePrefix').'pipelines', function (Blueprint $table) {
@@ -17,7 +16,7 @@ return new class extends Migration
 
         Schema::create(config('sales-management.tablePrefix').'pipeline_stages', function (Blueprint $table) {
             $table->id();
-            $table->integer("pipeline_id");
+            $table->unsignedBigInteger("pipeline_id");
             $table->string("name")->nullable();
             $table->string("description")->nullable();
             $table->string("color", 20)->nullable();
